+++ src/ua/user_identity_token.rs
@@ -2,9 +2,14 @@ use crate::ua;
 
 #[derive(Debug)]
 pub enum UserIdentityToken {
+    /// No user information is available.
     Anonymous(ua::AnonymousIdentityToken),
+    /// A user identified by user name and password.
     UserName(ua::UserNameIdentityToken),
+    /// A user identified by an X.509 v3 Certificate.
     X509(ua::X509IdentityToken),
+    /// A user identified by a token issued by an external authorization service.
+    Issued(ua::IssuedIdentityToken),
 }
 
 impl UserIdentityToken {
@@ -13,6 +18,7 @@ impl UserIdentityToken {
             Self::Anonymous(anonymous) => ua::ExtensionObject::new(anonymous),
             Self::UserName(user_name) => ua::ExtensionObject::new(user_name),
             Self::X509(x509) => ua::ExtensionObject::new(x509),
+            Self::Issued(issued) => ua::ExtensionObject::new(issued),
         }
     }
 }